apps/server/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
django-cors-headers==4.6.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
